+++ backend/src/main/java/com/carffeine/carffeine/filter/service/FilterService.java
@@ -1,15 +1,15 @@
 package com.carffeine.carffeine.filter.service;
 
-import com.carffeine.carffeine.filter.controller.domain.capacity.Capacity;
-import com.carffeine.carffeine.filter.controller.domain.capacity.CapacityRepository;
-import com.carffeine.carffeine.filter.controller.domain.companyName.CompanyName;
-import com.carffeine.carffeine.filter.controller.domain.companyName.CompanyNameRepository;
-import com.carffeine.carffeine.filter.controller.domain.connectorType.ConnectorType;
-import com.carffeine.carffeine.filter.controller.domain.connectorType.ConnectorTypeRepository;
 import com.carffeine.carffeine.filter.controller.dto.capacity.CapacitiesRequest;
 import com.carffeine.carffeine.filter.controller.dto.companyName.CompanyNamesRequest;
 import com.carffeine.carffeine.filter.controller.dto.connectorType.ConnectorTypesRequest;
 import com.carffeine.carffeine.filter.controller.dto.filter.FiltersResponse;
+import com.carffeine.carffeine.filter.domain.capacity.Capacity;
+import com.carffeine.carffeine.filter.domain.capacity.CapacityRepository;
+import com.carffeine.carffeine.filter.domain.companyName.CompanyName;
+import com.carffeine.carffeine.filter.domain.companyName.CompanyNameRepository;
+import com.carffeine.carffeine.filter.domain.connectorType.ConnectorType;
+import com.carffeine.carffeine.filter.domain.connectorType.ConnectorTypeRepository;
 import com.carffeine.carffeine.filter.exception.FilterException;
 import com.carffeine.carffeine.filter.exception.FilterExceptionType;
 import lombok.RequiredArgsConstructor;
